requirements.txt
google-cloud-storage
pandas
scikit-learn
pyarrow
gcsfs
google-cloud-bigquery
db-dtypes
dotenv
pandas-gbq
pendulum
pylance
hyperopt
xgboost
lightgbm
catboost
scipy
numpy
jinja2
# Dashboard + predictions API
streamlit
plotly
httpx
fastapi
uvicorn[standard]
pytest